readme.md

+++ step/orm/src/models.rs
@@ -1,6 +1,11 @@
-#[derive(Queryable)]
+
+
+use super::schema::posts;
+
+// 用于查询
+#[derive(Queryable, Debug)]
 pub struct Post {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
     pub body: String,
     pub published: bool,
